\name{archdata-package}
\alias{archdata-package}
\alias{archdata}
\docType{package}
\title{
archdata - Archaeological Data Sets
}
\description{
Archaeological data sets for use by archaeologists and others interested applying various methods of quantitative analysis using R.
}
\details{
\tabular{ll}{
Package: \tab archdata\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-06-24\cr
License: \tab GPL\cr
}
\describe{
  \item{\code{\link{Acheulean}}}{Seven African Acheulean Sites}
  \item{\code{\link{BarmoseI.grid}}}{Flakes per grid unit from Barmose I, South Zealand, Denmark}
  \item{\code{\link{BarmoseI.pp}}}{Piece plotted artifacts from Barmose I, South Zealand, Denmark}
  \item{\code{\link{DartPoints}}}{Five dart point types from Fort Hood, Texas, U.S.A.}
  \item{\code{\link{EndScrapers}}}{Upper Paleolithic End Scrapers from Castenet A and Ferrassie H, France}
  \item{\code{\link{EWBurials}}}{Ernest Witte Cemetery, Austin, County, Texas, U.S.A.}
  \item{\code{\link{Handaxes}}}{Lower Paleolithic handaxes from Furze Platt, Maidenhead, Berkshire, England}
  \item{\code{\link{Nelson}}}{Prehistoric Ceramics at Pueblo San Cristobal, New Mexico, U.S.A.}
  \item{\code{\link{Olorgesailie.maj}}}{Major stone tool classes, Olorgesailie, Kenya}
  \item{\code{\link{Olorgesailie.sub}}}{Stone tool subclasses, Olorgesailie, Kenya}
  \item{\code{\link{Snodgrass}}}{House pitss at the Mississippian Snodgrass site in Butler County, Missouri, U.S.A.}
}
}
\author{
David L. Carlson\cr
Maintainer: David L. Carlson <dcarlson@tamu.edu>\cr
}
\references{
See individual data sets for information on the source and publications illustrating their use.
}
\keyword{ package 
}
