\encoding{UTF-8}
\name{archdata-package}
\alias{archdata-package}
\alias{archdata}
\docType{package}
\title{
archdata - Archaeological Data Sets
}
\description{
Includes archaeological data sets used in \emph{Quantitative Methods in Archaeology Using R} by David L Carlson (Cambridge Manuals in Archaeology).
}
\details{
\tabular{ll}{
Package: \tab archdata\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2018-01-31\cr
License: \tab GPL\cr
}
\describe{
  \item{\code{\link{Acheulean}}}{Seven African Acheulean Sites}
  \item{\code{\link{Arnhofen}}}{Point pattern of mining pits from the Neolithic chert mine at Arnhofen}
  \item{\code{\link{BACups}}}{Bronze Age Cups from Italy}
  \item{\code{\link{BarmoseI.grid}}}{Flakes per grid unit from Barmose I, South Zealand, Denmark}
  \item{\code{\link{BarmoseI.pp}}}{Piece plotted artifacts from Barmose I, South Zealand, Denmark}
  \item{\code{\link{Bornholm}}}{Female Iron Age Graves, Bornholm, Denmark}
  \item{\code{\link{DartPoints}}}{Five dart point types from Fort Hood, Texas, U.S.A.}
  \item{\code{\link{EIAGraves}}}{Early Iron Age Graves - Tell el-Far'ah, Palestine}
  \item{\code{\link{EndScrapers}}}{Upper Paleolithic End Scrapers from Castenet A and Ferrassie H, France}
  \item{\code{\link{EngrBone}}}{Upper Paleolithic Engraved Bone Design Elements - Spain}
  \item{\code{\link{ESASites}}}{Early Stone Age Sites - Norway}
  \item{\code{\link{EWBurials}}}{Ernest Witte Cemetery, Austin, County, Texas, U.S.A.}
  \item{\code{\link{Fibulae}}}{Bronze La Tène fibulae from Műnsingen, Switzerland}
  \item{\code{\link{Handaxes}}}{Lower Paleolithic handaxes from Furze Platt, Maidenhead, Berkshire, England}
  \item{\code{\link{MaskSite}}}{Mask Site, Alaska, USA}
  \item{\code{\link{Mesolithic}}}{British Mesolithic assemblages}
  \item{\code{\link{Michelsberg}}}{Younger Neolithic Pottery from Central Europe}
  \item{\code{\link{Nelson}}}{Prehistoric Ceramics at Pueblo San Cristobal, New Mexico, U.S.A.}
  \item{\code{\link{Olorgesailie.maj}}}{Major stone tool classes, Olorgesailie, Kenya}
  \item{\code{\link{Olorgesailie.sub}}}{Stone tool subclasses, Olorgesailie, Kenya}
  \item{\code{\link{OxfordPots}}}{Distribution of Late Romano-British Oxford Pottery}
  \item{\code{\link{PitHouses}}}{Late Stone Age and Early Sami Iron Age Pithouses in Arctic Norway}
  \item{\code{\link{RBGlass1}}}{Romano-British Glass, Major and Minor Elements}
  \item{\code{\link{RBGlass2}}}{Romano-British Glass, Trace Elements}
  \item{\code{\link{RBPottery}}}{Romano-British Pottery}
  \item{\code{\link{Snodgrass}}}{House pits at the Mississippian Snodgrass site in Butler County, Missouri, U.S.A.}
  \item{\code{\link{TRBPottery}}}{Neolithic TRB Pottery from Demark}
}
}
\author{
David L. Carlson and Georg Roth\cr
Maintainer: David L. Carlson <dcarlson@tamu.edu>\cr
}
\references{
See individual data sets for information on the source and publications illustrating their use.
}
\keyword{ package
}
